Add utility to generate css text from data
Ref https://github.com/webstudio-is/webstudio-designer/issues/1019

We do a lot of changes with styles and ofte forget to keep saas up to
date. So instead whole logic to generate css will be hidden in shared
utility.
