+++ apps/designer/app/routes/s/css.ts
@@ -1,9 +1,11 @@
 import { type ActionArgs, json } from "@remix-run/node";
+import { generateCssText } from "@webstudio-is/project";
+import { db } from "@webstudio-is/project/server";
 import env from "~/env/env.public.server";
 import { getBuildParams } from "~/shared/router-utils";
 import { sentryException } from "~/shared/sentry";
-import { generateCssText } from "~/shared/css-utils";
 import { createContext } from "~/shared/context.server";
+import { loadCanvasData } from "~/shared/db";
 
 export const loader = async ({ request }: ActionArgs) => {
   try {
@@ -14,7 +16,29 @@ export const loader = async ({ request }: ActionArgs) => {
       throw json("Required project info", { status: 400 });
     }
 
-    const cssText = await generateCssText(buildParams, context);
+    const project = await db.project.loadByParams(buildParams, context);
+    if (project === null) {
+      throw json("Project not found", { status: 404 });
+    }
+
+    const canvasData = await loadCanvasData(
+      {
+        project,
+        env: buildParams.mode === "published" ? "prod" : "dev",
+        pageIdOrPath:
+          "pageId" in buildParams ? buildParams.pageId : buildParams.pagePath,
+      },
+      context
+    );
+    if (canvasData === undefined) {
+      throw json("Page not found", { status: 404 });
+    }
+
+    const cssText = generateCssText({
+      assets: canvasData.assets,
+      build: canvasData.build ?? undefined,
+      tree: canvasData.tree ?? undefined,
+    });
 
     return new Response(cssText, {
       headers: {
