--- apps/designer/app/shared/css-utils/index.ts
@@ -1,1 +1,0 @@
-export * from "./generate-css-text";
