+++ packages/project/src/shared/styles/index.ts
@@ -1,2 +1,3 @@
 export * from "./global-rules";
 export * from "./style-rules";
+export * from "./css";
